backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
whitenoise
httpx
tenacity
django-filter
